refactor: joga service.getPage pra dentro da CoeqDataTable, junto com os models relevantes
