+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/CoeqDataTable/model.ts
@@ -1,0 +1,13 @@
+import type {
+    ConditionIdentifiedBy,
+    PatientAgeRange,
+    PatientStatus,
+    VisitantCommunityHealthWorker,
+} from "@/features/acf/diabetes/common/model";
+
+export type SelectedFilterValuesCoeq = {
+    visitantCommunityHealthWorker: Array<VisitantCommunityHealthWorker>;
+    patientStatus: Array<PatientStatus>;
+    conditionIdentifiedBy: ConditionIdentifiedBy | "";
+    patientAgeRange: Array<PatientAgeRange>;
+};
