+++ src/features/acf/diabetes/frontend/model.ts
@@ -1,17 +1,6 @@
-import type {
-    ConditionIdentifiedBy,
-    PatientAgeRange,
-    PatientStatus,
-    VisitantCommunityHealthWorker,
-} from "../common/model";
+import type { SelectedFilterValuesCoeq } from "./AcfPage/modules/PanelSelector/modules/dashboards/modules/CoeqDataTable";
 
-export type SelectedFilterValuesCoeq = {
-    visitantCommunityHealthWorker: Array<VisitantCommunityHealthWorker>;
-    patientStatus: Array<PatientStatus>;
-    conditionIdentifiedBy: ConditionIdentifiedBy | "";
-    patientAgeRange: Array<PatientAgeRange>;
-};
-
+//TODO: Trocar pelo tipo real quando ele existir
 export type SelectedFilterValuesCoaps = SelectedFilterValuesCoeq & {
     campoDeCoaps: Array<string>;
 };
